+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log / Ray Tracing in One Weekend
 
 ### The Next Week
   - Fix    -- Remove premature source line for call to `get_sphere_uv` (#1701)
+  - Change -- Improve lifetime of float image data `fdata` in `rtw_image` class (#1723)
 
 ### The Rest of Your Life
 
@@ -19,7 +20,7 @@ Change Log / Ray Tracing in One Weekend
 ### Common
   - Fix    -- Fixed some dangling references to `random_in_unit_sphere()` (#1637)
   - Fix    -- Clarify `uniform_real_distribution` usage for `random_double()` (#1680)
-  - Update -- CMake minimum required version max now at 4.0.0.
+  - Change -- CMake minimum required version max now at 4.0.0.
 
 ### In One Weekend
   - Fix    -- Fix equation for refracted rays of non-unit length (#1644)